fix: Se modificó 'ft_split()' por compatibilidad

+++ ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: antgalan <antgalan@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/13 01:09:54 by edbander          #+#    #+#             */
-/*   Updated: 2023/01/30 21:30:17 by antgalan         ###   ########.fr       */
+/*   Updated: 2023/01/31 17:37:59 by antgalan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,7 +85,7 @@ static char	*_ft_printf_arg_hex(char type, va_list args)
 	if (type == 'p')
 	{
 		format = "0x";
-		longarg = (unsigned long long) va_arg(args, unsigned long);
+		longarg = va_arg(args, unsigned long long);
 		if (!longarg)
 			return ("0x0");
 		return (ft_itoahex_unsigned_long(
